+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Adding manual session control
   * Adding option to set tamper protection salt in config object
   * Adding calls to give and remove consent to all features at the same time
+  * Adding way to change the session update timer interval
   * Moving crash, event, view, rating related calls to a seperate modules
   * Deprecating functions that are being put into modules
   * Removing "addCrashLog" call